서준식/refactor 2  (#9)
* 수정사항 반영

* 컨벤션 적용
