+++ backend/src/main/java/com/WEB4_5_GPT_BE/dto/request/StudentSignUpRequest.java
@@ -1,2 +1,3 @@
-package com.WEB4_5_GPT_BE.dto.request;public record StudentSignUpRequest() {
-}
+package com.WEB4_5_GPT_BE.dto.request;
+
+public record StudentSignUpRequest() {}
