--- backend/src/main/java/com/WEB4_5_GPT_BE/unihub/domain/common/entity/Role.java
@@ -1,7 +1,0 @@
-package com.WEB4_5_GPT_BE.unihub.domain.common.entity;
-
-public enum Role {
-  STUDENT,
-  PROFESSOR,
-  ADMIN
-}
