+++ backend/src/main/java/com/WEB4_5_GPT_BE/unihub/domain/common/enums/DayOfWeek.java
@@ -1,4 +1,4 @@
-package com.WEB4_5_GPT_BE.unihub.domain.common.entity;
+package com.WEB4_5_GPT_BE.unihub.domain.common.enums;
 
 public enum DayOfWeek {
   MON,
