+++ backend/src/main/java/com/WEB4_5_GPT_BE/unihub/domain/common/enums/Role.java
@@ -1,0 +1,7 @@
+package com.WEB4_5_GPT_BE.unihub.domain.common.enums;
+
+public enum Role {
+  STUDENT,
+  PROFESSOR,
+  ADMIN
+}
